requirements.txt
streamlit
duckdb
pandas
pydantic
huggingface_hub<0.20
sentence-transformers
qdrant-client
openai
langchain
langchain-community
langchain-openai
PyMuPDF
neo4j
python-dotenv
plotly
watchdog
SQLAlchemy

